# arena alloc thrift buffers