Before:
```
BenchmarkTBinaryRT-8          	   10000	    134470 ns/op	   41601 B/op	     125 allocs/op
BenchmarkTCompactRT-8         	   10000	    131970 ns/op	   41426 B/op	     128 allocs/op
BenchmarkConcurrentRT-8       	   10000	    119215 ns/op	   47133 B/op	     310 allocs/op
BenchmarkHandlerUncompressed-8	  100000	     12828 ns/op	     544 B/op	      13 allocs/op
BenchmarkHandlerCompressed-8  	  100000	     23359 ns/op	     624 B/op	      18 allocs/op
```
After:
```
BenchmarkTBinaryRT-8          	   10000	    125916 ns/op	   39095 B/op	     117 allocs/op
BenchmarkTCompactRT-8         	   10000	    125230 ns/op	   39177 B/op	     121 allocs/op
BenchmarkConcurrentRT-8       	   10000	    119470 ns/op	   45342 B/op	     302 allocs/op
BenchmarkHandlerUncompressed-8	  100000	     12454 ns/op	     544 B/op	      13 allocs/op
BenchmarkHandlerCompressed-8  	  100000	     22764 ns/op	     624 B/op	      18 allocs/op
```